chore: Remove generated code from the CI
Addresses unwanted clutter in the PR that caused friction when reviewing things.
Add entry to .gitignore preventing accidentally checking them in again.

Note: it is expected to everyone who develops / deploys it to run RFB
themselves (e.g. `make gen`).

Kudos to @holzeis who highlighted the problem and attempted to solve it in