chore: Aligning imports
I am not entirely sure when to use cupertino, material or foundation yet. But the way I understood it, is we should only pick one. I guess it's not as important here, but wanted to prevent copy & paste variations.

We have also cupertino icons (ios) in our dependencies. Theoretically that should not be a problem, but we might want to make an active decision, which icons we'd like to use.