fix: Remove unnecessary Expanded widget
This fixes the error shown below. I guess it complains that the Expand widget tries to expand the containing widget to the full width, but can't do it since the SizedBoxed is fixed.

Incorrect use of ParentDataWidget.
The ParentDataWidget Expanded(flex: 1) wants to apply ParentData of type FlexParentData to a
RenderObject, which has been set up to accept ParentData of incompatible type ParentData.
Usually, this means that the Expanded widget has the wrong ancestor RenderObjectWidget. Typically,
Expanded widgets are placed directly inside Flex widgets.
The offending Expanded is currently placed inside a RepaintBoundary widget.