fix: Keep state in position tabs
The DefaultTabController seems to refresh the tab on every switch. This can be prevented by implementing the AutomaticKeepAliveClientMixin interface. Unfortunately it does not work on the CfdOffer Widget itself, but has to be applied to the tab view specifically. Hence I created a separate class. I am keeping it in one file though, as it logically tied to the CfdOffer.