feat: Add order detail screen
In the process of adding that feature I also reworked the state management for the cfd trading module. I introduced a stack to keep the current selection. This way the user can create a position, go back to her cfd overview check some positions and go back to the draft position without loosing data.