fix: Add internet permission
When testing the released apk on my android device I was not able to see the seed phrase, which indicates that the native rust lib has not been loaded. Pairing my android studio with my phone using wireless debug worked though. So I compared the two generated apks and noticied that the only difference is the added internet permission (which get automatically added in debug mode). Adding this permission also for the release version fixed that issue.