Split wallet into `Dashboard`, `Lightning` and `Bitcoin`
- `Dashboard` shows the tiles that guide the user, a combined view of balance and payment history and the quick-select service tiles. There are no action buttons in this view.
- `Lightning` is the view on the lightning wallet, it shows lightning specific payment history and an expandable FAB with `receive`/`send`
- `Bitcoin` is the view on the bitcoin wallet, it shows bitcoin specific payment (i.e. transaction) history and an expandable FAB with `deposit`/`withdraw`

The `Balance` widget was extended so it is aware of Bitcoin and Lightning balance and one can decide which to display based on a selector.
The payment model was extended so it covers Bitcoin specific history as well.

Note: Channel opening is not covered in FABs within the wallets at the moment, the way to open a channel will be the `Open Channel` tile in the `Dashboard` (for now, this can be changed/extended at a later point).