chore: Split initialisation and running of LDK
Without this, we are susceptible to a race condition (we need to *hope* that bdk
init will happen before we start other tasks depending on it). If we explicitly
wait for init first, we don't have such a problem.

Add `Clone` trait implementation to Wallet, so that we don't need to hold the mutex to the
wallet when passing some of its data around.
Note: cloning is pretty cheap here, because we're cloning only a bunch of Arc
pointers, not the underlying wallet data.

This makes the API more discoverable, allowing for the usage
`wallet.run_ldk(port)`.