feat: Add persistent SQLite database
I've used sqlx purely as we are familiar with it already. Otherwise, it's a bit of a pain
to use an async database here - a sync one would be much easier (due to frb
async boilerplate and runtime cost).

Database is managed for the lifetime of the app by the `ten_ten_one` library,
after being initialised with `db::init_db()` call.

Database in empty for now, but initialising and acquiring a connection works for
both maker and taker.

Note: There's no need to manually close connections to the DB, because we are
only storing a ConnectionPool, and the connections are closed with RAII when
they get out scope.