chore: Only run Android integration tests on main
Due to the high cost of running these tests (~30 min!), only run them periodically if
anything gets merged into main to ensure we can always release.

We're not using this job when deriving bors status, so we can avoid running it.
Hopefully this should make other job runners wait shorter.

We can always change this back after the tournament, but now bors gets in the
way a lot.