feat: Add margin and expiry calculation to generated order
The methods on the order will be exposed through the flutter rust bridge and can easily be used. e.g. that simplifies the method signatures as the required data is on the object / struct itself.

Note, this implementation has to be on the api level as otherwise it wouldn't be generated through frb. It would be nice to adapt the code generator target to target multiple rust files to allow for better code structuring.