feat: Support testnet
Revamp configuration to pass NETWORK env variable from command line (both for
taker and maker).
In case it's not set, or it's invalid, it defaults to testnet.

Move all the config values to a dedicated module to clear the code.

For local development, I recommend to set `export NETWORK=regtest` in your
bashrc to preserve previous behaviour.