fix: Remove channel peer data
I noticed that after combining the connection and run ldk functions, that the connection never got re-established after restarting. This was because after creating a channel the peer information is stored the channel_peer_data file. This channel_peer_data contained the peer information to which upon run_ldk the node tried to connect again.
I removed that part since
- we (already) manage connection centrally at another point and
- it's not a very resilient implementation - no error handling, etc.