Use `FormTextField` for quantity
Instead of a dropdown the user can specify a quantity between
We wrap this screen in a `Form` which is validated on editing the `TextFormField` as well as when hitting the submit button.

Note: In order to avoid running into display trouble on small screens a scrollable `ListView` was introduced. Potential warnings are not part of the `ListView` but pushed to the bottom below the `ListView`. This can still lead to a screen overflow when the user e.g. has the keyboard open to enter numbers, but the user should be able to see all relevant information.

The balance is not evaluated on purpose to let the user navigate to the `Confirm` page (where the `Confirm` button is then disabled).