feat: Add QR code scanner for lightning invoices
QR code scanner opens up in a separate widget. It fills in the lightning invoice
string if a valid QR code is found. Validation of the invoice is handled in the
send widget, as well as showing the details of the invoice.

Note: it should be fairly trivial to do the same for bitcoin QR codes, I will
proceed with it after this PR gets in.