Persist the channel opening `txid`
This enables us to properly ignore the maker's funding even when the channel was closed already.
We save the `open_channel_txid` upon the first time we are able to extract it from the channel. Once persisted we will always load it from the database for consecutive calls to sync the payment history.