Restrict tooltip to size of Bitcoin balance
The tooltip was created around the whole `Row`, which could cause trouble with opening the menu drawer in the `Bitcoin` wallet screen. The menu drawer was not properly clickable because the tooltip partly overlayed it.

Fixed by moving restricting the size of the `Row` to min (so it does not auto-expand) and moving the `Center` to the outside (it should not have been part of the `Balance` in the first place, the caller should decide how to layout the balance).