More resilient ignore logic
Deals with the possibility of previously failed channel opens that have an entry in the transactions to be ignored.
It's not trivial to associate the funding-txid of the maker with the actual channel-open-txid where the funding was used in our current flow.
This is hacky as fuck, but the whole solution of the maker funding the taker is hacky as fuck, so 🤷‍