Add onboarding tour on first startup
On first startup we show an onboarding tour including three screens:

1. one App all things Bitcoin: Welcome screen introducing 10101
2. your Keys, your Bitcoin: Explain that there is a Bitcoin and Lightning wallet that are created on first startup
3. Trade P2P non-custodial: List the three main steps of deposit, open channel, trade. Tell the user that the app will guide the user through these steps.

The onboarding tour is currently only shown on first startup, we persist a bool after that in the shared preferences. When the app is uninstalled and re-installed the onboarding tour is shown again (give that the app data was removed).