fix: Update on-chain balance immediately after channel closure
The coins are correctly moved to a pending state, like we do with any
other on-chain transaction.

The user is also informed about what to expect when closing the
channel.

It's important to note that if the user runs into a force-closure for
some other reason, the coins will also go into a pending state and be
reflected in the overall on-chain balance, but it will take _hundreds_
of confirmations for the coins to be confirmed.