Form validation for opening channel
Du to trouble using the `controller` for setting values in combination with the validation I opted for setting an initial value instead and adding a specific formatter.
The formatter combines well with the form validation - we should not just overwrite values that the user enters, but give the user an initial value and tell the user what is wrong if the form validation fails.
In order to get the initial value formatted we call the formatter specifically (unfortunately it is not called for the initial value out of the box).

Instead of only allowing channel open on a confirmed balance we allow it based on the external pending balance.
If the pending balance is not `0` we use the pending balance, otherwise the confirmed one to determine the maximum amount the taker can use.