Display `confirmed` Bitcoin balance instead of `total`
Displaying the total balance was a great effort, but it causes awkwardness because we then prominently display the maker amount being sent to the taker in the balance upon opening a channel.
We cannot easily overcome this problem in the current version, so it's better to display only the confirmed balance prominently.

Additionally this causes awkwardness when we run into a channel force close, because the balance is not `0`, but when trying to open a new channel we can only use `0`, because our displayed balance is actually not usable.