chore: Use async mutex for the wallet
With sync mutex, if a lock was held at the time another thread requested access
to the wallet, the request would return an error. This resulted in making it the
caller's responsibility to reschedule the request.

With an async mutex, the call will wait until the previous lock goes
out-of-scope and then acquire the lock.

This is not a problem, because we only acquire the lock to share a access to
underlying resources by calling `clone()`, therefore there should be no deadlocks.

The small price we pay for it is the overhead of `current_thread` tokio runtimes
in a few more places in the Rust API for flutter.