fix: Do not wrap Wallet in mutex
We are never actually mutating the `Wallet` after it has been
initialised, so it seems like we don't need a mutex to control access
to this resource.

This makes the code a bit simpler and maybe faster since we won't be
stuck waiting for some other function to release a mutex guard before
we can use the `Wallet`.